+++ config/routes.rb
@@ -1,2 +1,3 @@
 Txtblx::Engine.routes.draw do
+  resources :textblocks, only: [:update]
 end