+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Changes on `master`
 
+  * Route/Controller to update Textblocks
+
 ## Version  0.0.4
 
   * Removed automatic loading of Active Admin page.